cloud/proxy/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx[socks]>=0.27.0
python-multipart>=0.0.9
google-cloud-storage>=2.16.0
google-cloud-firestore>=2.16.0
firebase-admin>=6.5.0
